Seki_Web/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic